requirements.txt
alembic==1.8.1
fastapi==0.88.0
FastAPI-SQLAlchemy==0.2.1
psycopg2-binary==2.9.5
pydantic==1.10.2
SQLAlchemy==1.4.45
uvicorn==0.20.0
python-dotenv==0.21.0
bcrypt==4.0.1
cryptography==38.0.4
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.5
python-decouple==3.7
pytest==7.2.1
httpx==0.23.3